Change up how userlist.py's refresh interval is used ...
Have a min interval that is used initially and when things seem to have
gone wrong. From there, every time we callback "cleanly", double the last
interval used until we reach the target interval.

The motivation here is solving race conditions when we are initially
connecting to the server. Sometimes we don't seem to have all data
available yet because we aren't in all channels and/or weechat hasn't
done away checks for every yet. Thus this ... which maybe is overkill,
but probably isn't too much load on our client since we backoff.